requirements.txt
torch>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.1.0
ray[rllib]>=2.9.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
tensorboard>=2.15.0
pyyaml>=6.0
tqdm>=4.66.0
pytest>=7.4.0
